Update howto-mfa-nps-extension.md
https://learn.microsoft.com/en-us/azure/active-directory/authentication/howto-mfa-nps-extension-errors#next-steps

As described in the Troubleshooting Documentation, the user must have a license attributed so that it can authenticate with the 2nd authentication factor, the idea is to leave this information in a more explicit way in the prerequisites to avoid opening cases and generate doubt of those user.

NPS Extension for Azure AD MFA:
NPS Extension for Azure AD MFA only performs Secondary Auth for Radius requests in Access Accept State. Request received for User username with response state Access Reject, ignoring request.	

This error usually reflects an authentication failure in AD or that the NPS server is unable to receive responses from Azure AD. Verify that your firewalls are open bidirectionally for traffic to and from https://adnotifications.windowsazure.com and https://login.microsoftonline.com using ports 80 and 443. It is also important to check that on the DIAL-IN tab of Network Access Permissions, the setting is set to "control access through NPS Network Policy". This error can also trigger if the user is not assigned a license.